+++ src/test/java/seedu/mycrm/logic/commands/mails/AddTemplateCommandTest.java
@@ -243,7 +243,7 @@ class AddTemplateCommandTest {
         }
 
         @Override
-        public int getRevenue(LocalDate date) {
+        public double getRevenue(LocalDate date) {
             throw new AssertionError("This method should not be called.");
         }
