+++ src/test/java/seedu/mycrm/logic/commands/products/AddProductCommandTest.java
@@ -234,7 +234,7 @@ public class AddProductCommandTest {
         }
 
         @Override
-        public int getRevenue(LocalDate date) {
+        public double getRevenue(LocalDate date) {
             throw new AssertionError("This method should not be called.");
         }
         @Override
